requirements.txt
grpcio 
grpcio-tools 
kubernetes 
prometheus-api-client